Refactoring due to needed change in htaccess
Originally, htaccess was set to force php.ini in the projectbass folder
to apply to all folders.  This was done to facilitate relative paths for
require_once statements.  This, however, broke other sites on my server,
as well as PDO.  I restored htaccess to its original state, which means
that all relative paths need to be converted to absolute paths.